+++ packages/rrweb-snapshot/src/snapshot.ts
@@ -20,6 +20,8 @@ import {
   maskInputValue,
   isNativeShadowDom,
   getCssRulesString,
+  getInputType,
+  toLowerCase,
 } from './utils';
 
 let _id = 1;
@@ -31,12 +33,12 @@ export function genId(): number {
   return _id++;
 }
 
-function getValidTagName(element: HTMLElement): string {
+function getValidTagName(element: HTMLElement): Lowercase<string> {
   if (element instanceof HTMLFormElement) {
     return 'form';
   }
 
-  const processedTagName = element.tagName.toLowerCase().trim();
+  const processedTagName = toLowerCase(element.tagName);
 
   if (tagNameRegex.test(processedTagName)) {
     // if the tag name is odd and we cannot extract
@@ -221,8 +223,8 @@ function getHref() {
 
 export function transformAttribute(
   doc: Document,
-  tagName: string,
-  name: string,
+  tagName: Lowercase<string>,
+  name: Lowercase<string>,
   value: string | null,
 ): string | null {
   if (!value) {
@@ -268,20 +270,24 @@ export function _isBlockedElement(
   blockClass: string | RegExp,
   blockSelector: string | null,
 ): boolean {
-  if (typeof blockClass === 'string') {
-    if (element.classList.contains(blockClass)) {
-      return true;
-    }
-  } else {
-    for (let eIndex = element.classList.length; eIndex--; ) {
-      const className = element.classList[eIndex];
-      if (blockClass.test(className)) {
+  try {
+    if (typeof blockClass === 'string') {
+      if (element.classList.contains(blockClass)) {
         return true;
       }
+    } else {
+      for (let eIndex = element.classList.length; eIndex--; ) {
+        const className = element.classList[eIndex];
+        if (blockClass.test(className)) {
+          return true;
+        }
+      }
     }
-  }
-  if (blockSelector) {
-    return element.matches(blockSelector);
+    if (blockSelector) {
+      return element.matches(blockSelector);
+    }
+  } catch (e) {
+    //
   }
 
   return false;
@@ -313,22 +319,26 @@ export function needMaskingText(
   maskTextClass: string | RegExp,
   maskTextSelector: string | null,
 ): boolean {
-  const el: HTMLElement | null =
-    node.nodeType === node.ELEMENT_NODE
-      ? (node as HTMLElement)
-      : node.parentElement;
-  if (el === null) return false;
+  try {
+    const el: HTMLElement | null =
+      node.nodeType === node.ELEMENT_NODE
+        ? (node as HTMLElement)
+        : node.parentElement;
+    if (el === null) return false;
 
-  if (typeof maskTextClass === 'string') {
-    if (el.classList.contains(maskTextClass)) return true;
-    if (el.closest(`.${maskTextClass}`)) return true;
-  } else {
-    if (classMatchesRegex(el, maskTextClass, true)) return true;
-  }
+    if (typeof maskTextClass === 'string') {
+      if (el.classList.contains(maskTextClass)) return true;
+      if (el.closest(`.${maskTextClass}`)) return true;
+    } else {
+      if (classMatchesRegex(el, maskTextClass, true)) return true;
+    }
 
-  if (maskTextSelector) {
-    if (el.matches(maskTextSelector)) return true;
-    if (el.closest(maskTextSelector)) return true;
+    if (maskTextSelector) {
+      if (el.matches(maskTextSelector)) return true;
+      if (el.closest(maskTextSelector)) return true;
+    }
+  } catch (e) {
+    //
   }
   return false;
 }
@@ -629,7 +639,7 @@ function serializeElementNode(
       attributes[attr.name] = transformAttribute(
         doc,
         tagName,
-        attr.name,
+        toLowerCase(attr.name),
         attr.value,
       );
     }
@@ -674,12 +684,9 @@ function serializeElementNode(
       attributes.type !== 'button' &&
       value
     ) {
-      const type: string | null = n.hasAttribute('data-rr-is-password')
-        ? 'password'
-        : typeof attributes.type === 'string'
-        ? attributes.type.toLowerCase()
-        : null;
+      const type = getInputType(n);
       attributes.value = maskInputValue({
+        element: n,
         type,
         tagName,
         value,