+++ packages/rrweb/test/benchmark/dom-mutation.test.ts
@@ -30,6 +30,18 @@ const suites: Array<
     eval: 'window.workload()',
     times: 10,
   },
+  {
+    title: 'create 1000 DOM nodes and append into its previous looped node',
+    html: 'benchmark-dom-mutation-multiple-descendant-add.html',
+    eval: 'window.workload()',
+    times: 5,
+  },
+  {
+    title: 'create 10000 DOM nodes and move it to new container',
+    html: 'benchmark-dom-mutation-add-and-move.html',
+    eval: 'window.workload()',
+    times: 5,
+  },
 ];
 
 function avg(v: number[]): number {