+++ packages/rrdom-nodejs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # rrdom-nodejs
 
+## 2.0.0-alpha.8
+
+### Patch Changes
+
+- Updated dependencies [[`bc84246`](https://github.com/rrweb-io/rrweb/commit/bc84246f78849a80dbb8fe9b4e76117afcc5c3f7), [`d0fdc0f`](https://github.com/rrweb-io/rrweb/commit/d0fdc0f273bb156a1faab4782b40fbec8dccf915)]:
+  - rrweb-snapshot@2.0.0-alpha.8
+  - rrdom@2.0.0-alpha.8
+
 ## 2.0.0-alpha.7
 
 ### Patch Changes